Extract generic navigation file parsing functionality to own modules
So we can share as much of the navigation file parsing across different file types, we have extracted the core parsing functionality to separate modules.  This module is NavigationAidImporter.  It is included in RadioNavigationAidImporter.  The latter still defines using the class constant FIELDS the structure of each line.

Likewise, the storage of the loaded data is now moved to its separate module.  In this way, each navigation class has its own storage, but the way the data is organised and accessed is the same for all.