Implements improvements to the Aircraft class
Two factory methods created: #build which creates an aircraft with key performance set without a flightplan; #file_flightplan which sets the initial state of the aircraft based on the passed Flightplan.

The class is designed so that the state of the aircraft can be read by external objects.  However, once the aircraft has been created, only it can update its state.